migrating (most of) the remaining parameterized tests
