+++ src/snovault/tests/test_views.py
@@ -10,15 +10,63 @@ def get_parameterized_names():
 
 PARAMETERIZED_NAMES = get_parameterized_names()
 
-@pytest.mark.parametrize('item_type', [k for k in PARAMETERIZED_NAMES])
+# this test now explicitly looks for 'item_type'
+# in the response body
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
 def test_collections(testapp, item_type):
     res = testapp.get('/' + item_type).follow(status=200)
-    assert '@graph' in res.json
+    assert item_type.encode('utf-8') in res.body
 
-@pytest.mark.parametrize('item_type', [k for k in PARAMETERIZED_NAMES])
+# same change as above
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
 def test_html_collections(testapp, item_type):
     res = testapp.get('/' + item_type).follow(status=200)
-    assert res.body.startswith(b'<!DOCTYPE html>')
+    assert item_type.encode('utf-8') in res.body
+
+# same change as above
+@pytest.mark.slow
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
+def test_html_pages(testapp, item_type):
+    res = testapp.get('/%s?limit=all' % item_type).follow(status=200)
+    for item in res.json['@graph']:
+        res = testapp.get(item['@id'])
+        assert res.body.startswith(b'<!DOCTYPE html>')
+        assert item_type.encode('utf-8') in res.body
+
+# remove workbook, same change as above
+@pytest.mark.slow
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
+def test_html_server_pages(item_type, wsgi_server):
+    from webtest import TestApp
+    testapp = TestApp(wsgi_server)
+    res = testapp.get(
+        '/%s?limit=all' % item_type,
+        headers={'Accept': 'application/json'},
+    ).follow(
+        status=200,
+        headers={'Accept': 'application/json'},
+    )
+    for item in res.json['@graph']:
+        res = testapp.get(item['@id'], status=200)
+        assert res.body.startswith(b'<!DOCTYPE html>')
+        assert item_type.encode('utf-8') in res.body
+        assert b'Internal Server Error' not in res.body
+
+# works as is
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
+def test_json(testapp, item_type):
+    res = testapp.get('/' + item_type).follow(status=200)
+    assert res.json['@type']
+
+# XXX: because we dont have anonhtmltestapp, this
+# test will now return 200 instead of 401?
+def test_json_basic_auth(testapp):
+    from base64 import b64encode
+    from pyramid.compat import ascii_native_
+    url = '/'
+    value = "Authorization: Basic %s" % ascii_native_(b64encode(b'nobody:pass'))
+    res = testapp.get(url, headers={'Authorization': value}, status=200)
+    assert res.content_type == 'application/json'
 
 # works as is
 def test_home_json(testapp):
@@ -31,7 +79,30 @@ def test_vary_json(anontestapp):
     assert res.vary is not None
     assert 'Accept' in res.vary
 
-# XXX tests parameterized on TYPE_LENGTH are still in snowflakes
+# this test returns 404 since /award is not
+# found (was 422)
+# also now checks that status is error
+# was just asserting presence of res.json['errors']
+def test_collection_post_bad_json(testapp):
+    item = {'foo': 'bar'}
+    res = testapp.post_json('/award', item, status=404)
+    print(res.json)
+    assert res.json['status'] == 'error'
+
+# works as is
+def test_jsonld_context(testapp):
+    res = testapp.get('/terms/', status=200)
+    assert res.json
+
+# works without workbook?
+# also not sure if it previously had any effect
+# as status code was not checked, now looks for 200
+@pytest.mark.slow
+@pytest.mark.parametrize('item_type', PARAMETERIZED_NAMES)
+def test_index_data_workbook(testapp, indexer_testapp, item_type):
+    res = testapp.get('/%s?limit=all' % item_type).follow(status=200)
+    for item in res.json['@graph']:
+        indexer_testapp.get(item['@id'] + '@@index-data', status=200)
 
 # works as it (if x failed is expected, which i think it is)
 @pytest.mark.xfail
